requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.5.30
charset-normalizer==2.0.7
cloudinary==1.26.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.9.0
django-filter==21.1
django-heroku==0.3.1
django-on-heroku==1.1.2
django-rest-swagger==2.2.0
djangoflutterwave==0.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.8.6
psycopg2-binary==2.9.1
PyJWT==2.1.0
python-decouple==3.5
python-http-client==3.3.3
pytz==2021.3
requests==2.26.0
sendgrid==6.8.2
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
starkbank-ecdsa==1.1.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
